requirements.txt
# snowsearch
pyyaml==6.0.3
python-dotenv==1.2.1
aiohttp==3.13.2
brotli==1.2.0
requests==2.32.5
openai==2.8.1
tqdm==4.67.1
neo4j==6.0.3
tabulate==0.9.0
sentence-transformers==5.1.2
# grobid
grobid-client-python==0.1.2
grobid_tei_xml==0.1.3